Zadanie 4 - nie do końca działa
