+++ where_to_go/lib/dream_place_screen.dart
@@ -1,0 +1,104 @@
+import "package:flutter/material.dart";
+import "package:flutter_riverpod/flutter_riverpod.dart";
+import "features/favorite/favorite_provider.dart"; // ignore: unused_import
+import "features/models/dream_place.dart"; // ignore: unused_import
+import "features/places/places_provider.dart";
+
+class DreamPlaceScreen extends ConsumerWidget {
+  static const route = "/place";
+  final String id;
+
+  const DreamPlaceScreen({super.key, required this.id});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final places = ref.watch(placesProvider);
+    final place = places.firstWhere((p) => p.id == id);
+
+    return Scaffold(
+      backgroundColor: Colors.pink[300],
+      appBar: AppBar(
+        title: Text(place.title),
+        actions: [
+          IconButton(
+            icon: Icon(
+              place.isFavorite ? Icons.favorite : Icons.favorite_border,
+              color: place.isFavorite ? Colors.red : null,
+            ),
+            onPressed: () {
+              ref.read(placesProvider.notifier).toggleFavorite(id);
+            },
+          ),
+        ],
+      ),
+      body: SingleChildScrollView(
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(12),
+              child: DecoratedBox(
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(12),
+                  boxShadow: const [
+                    BoxShadow(
+                      color: Colors.black26,
+                      blurRadius: 8,
+                      offset: Offset(0, 4),
+                    )
+                  ],
+                ),
+                child: Hero(
+                  tag: place.title,
+                  child: ClipRRect(
+                    borderRadius: BorderRadius.circular(12),
+                    child: Image.asset(
+                      place.imagePath,
+                      fit: BoxFit.cover,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    place.placeName,
+                    style: const TextStyle(
+                      fontSize: 26,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  const SizedBox(height: 8),
+                  Text(
+                    place.description,
+                    style: const TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.w600,
+                      height: 1.4,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            const SizedBox(height: 16),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: place.attractions.map((attr) {
+                return Column(
+                  children: [
+                    Icon(attr.icon, size: 40),
+                    Text(attr.label),
+                  ],
+                );
+              }).toList(),
+            ),
+            const SizedBox(height: 16),
+          ],
+        ),
+      ),
+    );
+  }
+}
