+++ where_to_go/lib/features/models/attraction.dart
@@ -1,0 +1,8 @@
+import "package:flutter/material.dart";
+
+class Attraction {
+  final IconData icon;
+  final String label;
+
+  const Attraction({required this.icon, required this.label});
+}
