+++ where_to_go/pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   hooks_riverpod: ^2.6.1
   riverpod_annotation: ^2.6.1
   flutter_riverpod: ^2.6.1
+  go_router: ^16.2.0
 
 dev_dependencies:
   flutter_test:
